+++ CHANGELOG.md
@@ -74,6 +74,8 @@
 - InputText: Fixed missing suggestions in input fields.
 - SearchInterfaceOsmNominatim: Fixed getter path in debounce check
 - GroupLayer: Fixed errors when opening layer info for layers within a group.
+- wfsSearch: secured layers can now be searched.
+- Issue #1410: Legend – Fixed ordering of legend entries to match the style configuration.
 
 ---